move Teacup.calculate into its own file, to resolve dependency detection issues
